move graphdb
