+++ pipelines/matrix/conf/base/moa_extraction/parameters.yml
@@ -1,14 +1,14 @@
 moa_extraction:
   # Note: We need to use the same database name as is set in catalog.yml
   gdb_two_hop: 
-    object: matrix.pipelines.embeddings.nodes.GraphDB
+    object: matrix.pipelines.moa_extraction.utils.GraphDB
     endpoint: ${globals:neo4j.host}
     auth:
       -  ${globals:neo4j.user}
       -  ${globals:neo4j.password}
     database: analytics-two-hop
   gdb_three_hop:
-    object: matrix.pipelines.embeddings.nodes.GraphDB
+    object: matrix.pipelines.moa_extraction.utils.GraphDB
     endpoint: ${globals:neo4j.host}
     auth:
       -  ${globals:neo4j.user}
